pyproject.toml
[project]
name = "wedding-bot-api"
version = "0.1.0"
description = "RAG-powered wedding assistant API"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.9.0",
    "azure-identity>=1.19.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]
